refactor sidekiq workers out, puts them in PsulibTraject module
