+++ Dockerfile
@@ -12,8 +12,8 @@ RUN mkdir /app/tmp
 RUN chown -R app /app
 USER app
 
-RUN gem install bundler -v "$(grep -A 1 "BUNDLED WITH" Gemfile.lock | tail -n 1)"
 COPY --chown=app Gemfile Gemfile.lock /app/
+RUN gem install bundler -v "$(grep -A 1 "BUNDLED WITH" Gemfile.lock | tail -n 1)"
 RUN bundle install
 
 COPY . /app
