+++ config/sidekiq.yml
@@ -6,5 +6,5 @@
 :schedule:
   hourlies:
     cron: '3 0,7-23 * * *'
-    class: 'HourliesWorker'
+    class: 'PsulibTraject::HourliesWorker'
