+++ lib/psulib_traject.rb
@@ -13,6 +13,9 @@ require 'yaml'
 module PsulibTraject
   require 'psulib_traject/hathi_overlap_reducer'
   require 'psulib_traject/indexer'
+  require 'psulib_traject/index_file_worker'
+  require 'psulib_traject/index_worker'
+  require 'psulib_traject/hourlies_worker'
   require 'psulib_traject/macros'
   require 'psulib_traject/marc_combining_reader'
   require 'psulib_traject/processors/access_facet'
